Query VCF

# Alternative methods #### Rsamtools #### tab <- Rsamtools::scanTabix(file = vcf_url, gr) #### Rtracklayer #### tab <- Rsamtools::TabixFile(vcf_url) open(tab) tab <- rtracklayer::import(tab, which=gr) #### seqminer #### tab <- seqminer::tabix.read(tabixFile = vcf_url, tabixRange = coord_range) header <- VariantAnnotation::scanVcfHeader(file = path)

query_vcf_variantannotation(
  vcf_url,
  dat,
  samples = character(),
  genome = "GRCh37",
  save_path = tempfile(fileext = ".vcf.bgz"),
  verbose = TRUE
)

Arguments

vcf_url
URL or path to VCF.

dat
SNP-level data table.

samples
Sample names to subset the VCF by before computing LD.

genome
A character or Seqinfo object.

character: Genome identifier as a single string or named character vector. Names of the character vector correspond to chromosome names in the file. This identifier replaces the genome information in the VCF Seqinfo (i.e., seqinfo(vcf)). When not provided, genome is taken from the VCF file header.

Seqinfo: When genome is provided as a Seqinfo it is propagated to the VCF output. If seqinfo information can be obtained from the file, (i.e., seqinfo(scanVcfHeader(fl)) is not empty), the output Seqinfo is a product of merging the two. If a param (i.e., ScanVcfParam) is used in the call to readVcf, the seqlevels of the param ranges must be present in genome.

verbose
Print messages.

Source

VariantAnnotation filtering

Examples

if (FALSE) {
BST1 <- echodata::BST1
dat <- BST1[seq(1, 50), ]
vcf_url <- file.path(
    "ftp://ftp-trace.ncbi.nih.gov/1000genomes/ftp/release/20110521/",
    "ALL.chr4.phase1_release_v3.20101123.snps_indels_svs.genotypes.vcf.gz"
)
vcf <- query_vcf_variantannotation(
    vcf_url = vcf_url,
    dat = dat
)
}